Implement codesplitting and use types from package
